confirmation for destroying nodes fixes #219

+++ saltcloud/cli.py
@@ -95,8 +95,13 @@ class SaltCloud(parsers.SaltCloudParser):
             else:
                 names = self.config.get('names', None)
 
+            msg = 'The following virtual machines are set to be destroyed:\n'
+            for name in names:
+                msg += '  {0}\n'.format(name)
+
             try:
-                mapper.destroy(names)
+                if self.print_confirm(msg):
+                    mapper.destroy(names)
             except Exception as exc:
                 print('There was an error: {0}'.format(exc))
             self.exit(0)
@@ -108,8 +113,13 @@ class SaltCloud(parsers.SaltCloudParser):
             else:
                 names = self.config.get('names', None)
 
+            msg = 'The following virtual machines are set to be actioned with "{0}":\n'.format(self.options.action)
+            for name in names:
+                msg += '  {0}\n'.format(name)
+
             try:
-                mapper.do_action(names)
+                if self.print_confirm(msg):
+                    mapper.do_action(names)
             except Exception as exc:
                 print('There was an error: {0}'.format(exc))
             self.exit(0)
@@ -126,7 +136,27 @@ class SaltCloud(parsers.SaltCloudParser):
                 print('Nothing to do')
                 self.exit(0)
             try:
-                mapper.run_map()
+                dmap = self.map_data()
+
+                msg = 'The following virtual machines are set to be created:\n'
+                for name in dmap['create']:
+                    msg += '  {0}\n'.format(name)
+                if 'destroy' in dmap:
+                    msg += 'The following virtual machines are set to be destroyed:\n'
+                    for name in dmap['destroy']:
+                        msg += '  {0}\n'.format(name)
+
+                if self.print_confirm(msg):
+                    mapper.run_map(dmap)
             except Exception as exc:
                 print('There was an error: {0}'.format(exc))
             self.exit(0)
+
+
+    def print_confirm(self, msg):
+        print(msg)
+        res = raw_input('Proceed? [N/y]')
+        if not res.lower().startswith('y'):
+            return False
+        print('...proceeding')
+        return True
